--- src/invest/invest-migrate/index.ts
@@ -1,1 +1,0 @@
-export * from './invest-migrate'
